refactor(building_blocks): move pagination contract to domain layer
Relocate the repository pagination contract (Pagination,
PaginatedResult) from building_blocks/application/pagination to a new
building_blocks/domain/pagination module so domain repository
interfaces no longer depend on an application-layer type for their own
signatures (inverted dependency, clean-arch M-4/B-11).

The cursor codec and limit-clamping constants stay in the application
module — they are mechanisms consumed by application, infrastructure,
and presentation, not part of the domain contract.
