+++ src/building_blocks/application/pagination.py
@@ -29,6 +29,12 @@ point instead of a 400 in the middle of an infinite scroll.
 This building block is intentionally tiny and framework-agnostic so
 both the application layer (DTOs, use cases) and the infrastructure
 layer (SQLAlchemy queries) can share the same shapes.
+
+The repository pagination *contract* (``Pagination`` and
+``PaginatedResult``) lives in ``building_blocks/domain/pagination.py``
+so domain repository interfaces can declare it without depending on the
+application layer; this module holds the cursor codec and the ``limit``
+clamping constants that drive those repositories.
 """
 
 from __future__ import annotations
@@ -36,9 +42,6 @@ from __future__ import annotations
 import base64
 import binascii
 from dataclasses import dataclass
-from typing import Generic, TypeVar
-
-T = TypeVar("T")
 
 # Default and clamp values for the `limit` query param. Routes typically
 # expose `limit` as an int and clamp to ``[1, MAX_PAGE_SIZE]`` before
@@ -84,48 +87,6 @@ class DualCursorValue:
 
     movies: str | None
     series: str | None
-
-
-@dataclass(frozen=True)
-class Pagination:
-    """Pagination metadata returned alongside a page of items.
-
-    Attributes:
-        next_cursor: Opaque token to pass back as `cursor` on the next
-            request. ``None`` when there are no more pages.
-        has_more: Convenience flag — equivalent to ``next_cursor is not
-            None`` but explicit so clients don't have to infer it.
-    """
-
-    next_cursor: str | None
-    has_more: bool
-
-
-@dataclass(frozen=True)
-class PaginatedResult(Generic[T]):
-    """A page of results with its pagination metadata.
-
-    Attributes:
-        items: The page's items.
-        pagination: ``Pagination`` for ``next_cursor`` / ``has_more``.
-        total_count: Total rows matching the query, or ``None`` when
-            the caller did not request it. Computing the total requires
-            an extra ``COUNT(*)`` query, so it's opt-in.
-        item_cursors: Parallel to ``items`` — cursor that resumes
-            strictly after that item. Populated only by repository
-            methods whose consumers need to advance through a partial
-            prefix of the page (e.g. the catalog "by genre" listing,
-            which fetches movies and series in parallel and may use
-            only some items from each before the merged page is full).
-            For straight-through consumers that always exhaust the
-            whole page, ``pagination.next_cursor`` is enough and this
-            field stays ``None``.
-    """
-
-    items: list[T]
-    pagination: Pagination
-    total_count: int | None = None
-    item_cursors: list[str] | None = None
 
 
 def encode_cursor(internal_id: int) -> str:
@@ -266,8 +227,6 @@ __all__ = [
     "MAX_PAGE_SIZE",
     "CursorValue",
     "DualCursorValue",
-    "Pagination",
-    "PaginatedResult",
     "TitleCursorValue",
     "decode_cursor",
     "decode_dual_cursor",
